Add files via upload
In this project, hypotheses related to the musical preferences of two cities were tested. This required analyzing data from an actual streaming service to compare the behavior of users from Springfield and Shelbyville. The project was divided into three stages:

An overview of the data and observations was provided.
Pre-processing of the data, including filtering and cleaning.
Hypothesis testing by performing the necessary programming to test each outcom