Add files via upload
In this project, we received real data from Instacart, a grocery delivery platform where customers can place a grocery order and have their shopping delivered to their home. The provided data were taken from a Kaggle competition. It was necessary to clean the data and prepare a report that provided insights into the shopping habits of Instacart customers.